+++ spark-bigquery-dsv2/spark-3.5-bigquery-lib/src/main/java/com/google/cloud/spark/bigquery/BigQueryCatalog.java
@@ -17,7 +17,9 @@ package com.google.cloud.spark.bigquery;
 
 import com.google.api.client.googleapis.json.GoogleJsonResponseException;
 import com.google.cloud.bigquery.BigQueryException;
+import com.google.cloud.bigquery.Dataset;
 import com.google.cloud.bigquery.DatasetId;
+import com.google.cloud.bigquery.DatasetInfo;
 import com.google.cloud.bigquery.Schema;
 import com.google.cloud.bigquery.TableDefinition;
 import com.google.cloud.bigquery.TableId;
@@ -33,9 +35,9 @@ import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
 import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Streams;
 import com.google.inject.Injector;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
 import java.util.ServiceLoader;
@@ -142,13 +144,28 @@ public class BigQueryCatalog implements TableCatalog, SupportsNamespaces {
           () ->
               // TODO: reuse injector
               Spark3Util.createBigQueryTableInstance(
-                  Spark35BigQueryTable::new,
-                  null,
-                  ImmutableMap.of(
-                      "dataset", identifier.namespace()[0], "table", identifier.name())));
+                  Spark35BigQueryTable::new, null, toLoadProperties(identifier)));
     } catch (ExecutionException e) {
-      throw new BigQueryConnectorException("Problem loaing table " + identifier, e);
+      throw new BigQueryConnectorException("Problem loading table " + identifier, e);
     }
+  }
+
+  Map<String, String> toLoadProperties(Identifier identifier) {
+    ImmutableMap.Builder<String, String> result = ImmutableMap.builder();
+    switch (identifier.namespace().length) {
+      case 1:
+        result.put("dataset", identifier.namespace()[0]);
+        break;
+      case 2:
+        result.put("project", identifier.namespace()[0]);
+        result.put("dataset", identifier.namespace()[1]);
+        break;
+      default:
+        throw new IllegalArgumentException(
+            "The identifier [" + identifier + "] is not recognized by BigQuery");
+    }
+    result.put("table", identifier.name());
+    return result.build();
   }
 
   @Override
@@ -268,12 +285,41 @@ public class BigQueryCatalog implements TableCatalog, SupportsNamespaces {
 
   @Override
   public String[][] listNamespaces() throws NoSuchNamespaceException {
-    return new String[0][];
+    return Streams.stream(bigQueryClient.listDatasets())
+        .map(Dataset::getDatasetId)
+        .map(this::toNamespace)
+        .toArray(String[][]::new);
   }
 
+  private String[] toNamespace(DatasetId datasetId) {
+    return new String[] {datasetId.getDataset()};
+  }
+
+  private DatasetId toDatasetId(String[] namespace) {
+    switch (namespace.length) {
+      case 1:
+        return DatasetId.of(namespace[0]);
+      case 2:
+        return DatasetId.of(namespace[0], namespace[1]);
+      default:
+        throw new IllegalArgumentException(
+            "The identifier [" + Arrays.toString(namespace) + "] is not recognized by BigQuery");
+    }
+  }
+
+  // based on JDBCTableCatalog.listNamespaces(namespace)
   @Override
   public String[][] listNamespaces(String[] namespace) throws NoSuchNamespaceException {
-    return new String[0][];
+    switch (namespace.length) {
+      case 0:
+        return listNamespaces();
+      case 1:
+        if (namespaceExists(namespace)) {
+          return new String[0][];
+        }
+    }
+    // unsupported case
+    throw new NoSuchNamespaceException(namespace);
   }
 
   @Override
@@ -281,19 +327,32 @@ public class BigQueryCatalog implements TableCatalog, SupportsNamespaces {
     return datasetExistenceCache.getUnchecked(namespace);
   }
 
-  private boolean bigQueryDatasetExists(String[] dataset) {
-    return bigQueryClient.datasetExists(DatasetId.of(dataset[0]));
+  private boolean bigQueryDatasetExists(String[] namespace) {
+    return bigQueryClient.datasetExists(toDatasetId(namespace));
   }
 
   @Override
   public Map<String, String> loadNamespaceMetadata(String[] namespace)
       throws NoSuchNamespaceException {
-    return Collections.emptyMap();
+    DatasetInfo dataset = bigQueryClient.getDataset(toDatasetId(namespace));
+    ImmutableMap.Builder<String, String> result = ImmutableMap.builder();
+    result.put("bigquery_location", dataset.getLocation());
+    result.put("creationTime", dataset.getCreationTime().toString());
+    Optional.ofNullable(dataset.getDescription())
+        .ifPresent(description -> result.put("comment", description));
+    return result.build();
   }
 
   @Override
   public void createNamespace(String[] namespace, Map<String, String> metadata)
-      throws NamespaceAlreadyExistsException {}
+      throws NamespaceAlreadyExistsException {
+    Preconditions.checkArgument(
+        namespace.length == 1, "BigQuery does not support nested namespaces");
+    if (namespaceExists(namespace)) {
+      throw new NamespaceAlreadyExistsException(namespace);
+    }
+    bigQueryClient.createDataset(DatasetId.of(namespace[0]), metadata);
+  }
 
   @Override
   public void alterNamespace(String[] namespace, NamespaceChange... changes)
@@ -302,6 +361,12 @@ public class BigQueryCatalog implements TableCatalog, SupportsNamespaces {
   @Override
   public boolean dropNamespace(String[] namespace, boolean cascade)
       throws NoSuchNamespaceException, NonEmptyNamespaceException {
-    return false;
+    if (!namespaceExists(namespace)) {
+      throw new NoSuchNamespaceException(namespace);
+    }
+    if (!cascade && listTables(namespace).length > 0) {
+      throw new NonEmptyNamespaceException(namespace);
+    }
+    return bigQueryClient.deleteDataset(toDatasetId(namespace), cascade);
   }
 }