+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## Next
 * Added new connector, `spark-4.0-bigquery` aimed to be used in Spark 4.0. Like Spark 4.0, this connector requires at
   least Java 17 runtime. It is currently in preview mode.
+* PR #1367: Query Pushdown is no longer supported.
+* PR #1369: Catalog enhancements
 * PR #1370: Add backwards compatibility for OpenLineage prior to version 1.18.0.
 
 ## 0.42.2 - 2025-05-16